Unify implementations of TryAdd and Assoc in binary search trees.
Split up the association of a key and value in a binary search tree into
two steps - one to locate the affected node and the type of update to
perform, and one to apply this update to the tree. This means that
Assoc can reject the update if it would modify an existing node since
the key already exists.