Combine add and assoc operations.
Update recursive Assoc and RootAssoc operations in RandomisedBinarySearchTree
to account for the case when a key is being added. If a new key-value
pair is being added, do not update the node associated with the existing key.

Update TryAdd to use this method so it no longer performs two searches
through the tree for the key.