Add red-black tree implementation.
Add implementation for red-black tree. The nodes in the tree have an
explicit reference to their parents so add an IHasParent<T> interface
for types with a parent of type T. The ability to access the parent
directly simplifies some tree operations so add implementations for
rotation etc. which make use of this. Rename the existing RotateLeft
and RotateRight methods to RotateLeftFromParent and RotateRightFromParent.