Extract search logic in binary search trees.
Some binary search tree operations perform an initial search
for the key and find the position it would occupy in the tree
if it does not exist e.g. Insert and Delete operations. Some
variants also need to modify the structure of the tree after
the operation to maintain balance invariants.

Extract the search logic which returns an object containing
the status of the search (successful or not) and the search
path taken through the tree to find the key.

Refactor BinarySearchTree to use this extracted version.